Update LogWithTernary.java
Add Java program demonstrating the use of ternary operator in authentication

This commit adds a small Java program, LogWithTernary.java, that showcases the usage of the ternary operator in the first step of authentication. The program takes user input for login and uses the ternary operator to determine the role (Administrator or User), which is then concatenated to a greeting string.